requirements.txt
streamlit
requests
g4f[all]
matplotlib
Flask
python-dotenv
requests